Only auto-merge js/json files in the current folder
Trying to auto-merge all nested folder modules is "clever", but comes with too many headaches.

For example if the script installs another node.js project in a subfolder, that project may have all kinds of js/json files.

By auto-merging all the js/json files in the nested folder, the root object will have all the useless subtrees and sometimes this results in crashes.

By only limiting the auto-merge to the current folder, it's not as "powerful", but still can get as much done just by putting all the required files in the same folder.