Path order is important
Generic package managers like conda and homebrew should come later than things like python and node, because homebrew and conda can install python as a dependency of some other package, and that will overwrite the python/node versions we want. To ensure that we are using the exact version for python/node, they need to come first.