env api
## 1. env.set

set the environment variable of the currently running app (updates the ENVIRONMENT file at the root of the app repository)

syntax similarto `local.set`. just key/value

## 2. env.switch

convenience method to switch through multiple options when setting the ENVIRONMENT file.

```
{
  method: "env.switch",
  params: {
    <key>: <options>
  }
}
```

the `<key>` will be set to a value in the `<options>` array.

- When the key doesn't exist, the default item (the first item in the options array) will be set to the value.
- When the key exists, check for the existing value, and update the value to the next item in the `options` array.