Save traffic
previously the backend was sending everything to the frontend via websocket, but this wastes too much traffic.

it's ok when just running on localhost, but this is no longer the case so need to save traffic. only sending the "raw" tokens instead of the full terminal state when running scripts