Fix terminal when running in a background tab
the terminal doesn't print to the screen when new tokens are written while the tab is in the background. Even when it comes back to foreground it doesn't update. Might be because the terms are inside iframes.

For now, the solution is to add a new line when coming back from background.

But can't keep adding new lines every time i switch tabs, since it will keep adding new lines. So add a "dirty" attribute to only add a new line when there have been some updates (dirty: true)