Prioritizing Visual Studio Build Tools Environment variables
When a conda environment is activated, the activation script makes all the conda related environment variables top priority. This causes it to find incorrect binaries when trying to build stuff. For example instead of using the `link.exe` in the ms build tools, it tries to find `link.exe` in the conda environment, and in some cases the `link.exe` does exist from some conda package. This is not the desired behavior since we need to use ms build tools on windows. So as an exception, prioritize build tools related environment variables by:

1. NOT setting them before the conda activation
2. and PREPENDING them AFTER the conda activation