Corrupt Visual Studio Build Tools Fix
Sometimes the user has a corrupt (or paused) installation of VS build tools, which does nothing when triggered by a launcher. Instead of installing from fresh or resuming, it just returns immediately if the installs are in those states.

To avoid this, just try uninstall first, and then install for all cases.

UX wise it's a bit more clunky but not that bad. Ensures that every install is fresh.